refactor(test): Extract out `readFixtureSync` to own utility file
