+++ tests/test_no_syntax_error.js
@@ -1,24 +1,10 @@
 'use strict';
 
-const fs = require('fs');
-const path = require('path');
 const util = require('util');
 const Parser = require('../lib/parsing.js');
+const {readFixtureSync} = require('./utils.js');
 
-/**
- * @typedef {object} FixtureEntry
- * @property {boolean} skip
- * @property {string} typeExprStr
- * @property {FilePosition} position
- *
- * @typedef {object} FilePosition
- * @property {string} fileName
- * @property {string} lineno
- *
- * @typedef {{fileName: string, filePath: string} & FixtureEntry[]} Fixture
- */
-
-/** @type {{[fixtureName: string]: Fixture}} */
+/** @type {{[fixtureName: string]: import('./utils').Fixture}} */
 const Fixtures = {
   CATHARSIS: readFixtureSync('catharsis-types'),
   CLOSURE_LIBRARY: readFixtureSync('closure-library-types'),
@@ -31,8 +17,6 @@ describe('Parser', function() {
   Object.keys(Fixtures).forEach(function(fixtureName) {
     const fixture = Fixtures[fixtureName];
     it(`should not throw any errors when parsing ${fixture.filePath}`, function() {
-      /** @type {Error[]} */
-      let errors = [];
       fixture.forEach(function({skip, typeExprStr, position}) {
         if (skip) return;
 
@@ -54,44 +38,3 @@ describe('Parser', function() {
     });
   });
 });
-
-/**
- * @param {string} fileName
- * @return {Fixture}
- */
-function readFixtureSync(fileName) {
-  const filePath = path.resolve(__dirname, 'fixtures', fileName);
-
-  /** @type {any} */
-  let result = fs.readFileSync(filePath, 'utf8')
-    .trim()
-    .split(/\n/)
-    .map(function(line, lineIdx) {
-      return {
-        // When the line starts with "//", we should skip it.
-        skip: /^\/\//.test(line),
-
-        typeExprStr: line.trim().replace(/^\{(.*)\}$/, '$1').replace(/\\n/g, '\n'),
-        position: {
-          fileName,
-          lineno: lineIdx + 1,
-        },
-      };
-    });
-
-  Object.defineProperties(result, {
-    fileName: {
-      value: fileName,
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    },
-    filePath: {
-      value: path.relative(process.cwd(), filePath),
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    },
-  });
-  return result;
-}
